+++ src/app/cms/cms.module.ts
@@ -12,21 +12,21 @@ import { StoreModule } from '@ngrx/store'
 
 import { environment } from '../../environments/environment'
 import { CmsRoutingModule } from './cms-routing.module'
-import { CmsComponent } from './components/cms-root/cms.component'
-import { HeaderComponent } from './components/common/header/header.component'
-import { ImageSelectorComponent } from './components/common/image-selector/image-selector.component'
-import { ImageComponent } from './components/common/image/image.component'
-import { InputFileComponent } from './components/common/input-file/input-file.component'
-import { PictureGridComponent } from './components/common/picture-grid/picture-grid.component'
-import { SidenavListComponent } from './components/common/sidenav-list/sidenav-list.component'
-import { DashboardComponent } from './components/routes/dashboard/dashboard.component'
-import { ProgressCardComponent } from './components/routes/file-upload/components/progress-card/progress-card.component'
-import { UploadCardComponent } from './components/routes/file-upload/components/upload-card/upload-card.component'
-import { FileUploadComponent } from './components/routes/file-upload/file-upload.component'
-import { StorageService } from './components/routes/file-upload/service/storage.service'
-import { UploadStateService } from './components/routes/file-upload/service/upload-state.service'
-import { LoginComponent } from './components/routes/login/login.component'
-import { PublishComponent } from './components/routes/publish/publish.component'
+import { CmsComponent } from './root/cms.component'
+import { HeaderComponent } from './components/header/header.component'
+import { ImageSelectorComponent } from './components/image-selector/image-selector.component'
+import { ImageComponent } from './components/image/image.component'
+import { InputFileComponent } from './components/input-file/input-file.component'
+import { PictureGridComponent } from './components/picture-grid/picture-grid.component'
+import { SidenavListComponent } from './components/sidenav-list/sidenav-list.component'
+import { DashboardComponent } from './routes/dashboard/dashboard.component'
+import { ProgressCardComponent } from './routes/file-upload/components/progress-card/progress-card.component'
+import { UploadCardComponent } from './routes/file-upload/components/upload-card/upload-card.component'
+import { FileUploadComponent } from './routes/file-upload/file-upload.component'
+import { StorageService } from './routes/file-upload/service/storage.service'
+import { UploadStateService } from './routes/file-upload/service/upload-state.service'
+import { LoginComponent } from './routes/login/login.component'
+import { PublishComponent } from './routes/publish/publish.component'
 import { MaterialModule } from './material/material.module'
 import { AuthService } from './services/auth.service'
 import { DatabaseService } from './services/database.service'
