+++ src/app/cms/components/header/header.component.ts
@@ -1,7 +1,7 @@
 import { Component, EventEmitter, OnInit, Output } from '@angular/core'
-import { AuthService } from '../../../services/auth.service'
+import { AuthService } from '../../services/auth.service'
 import { Store } from '@ngrx/store'
-import * as fromCms from '../../../state/cms.reducer'
+import * as fromCms from '../../state/cms.reducer'
 import { Observable } from 'rxjs'
 
 @Component({
