+++ src/app/cms/components/image/image.component.scss
