+++ src/app/cms/components/sidenav-list/sidenav-list.component.scss
