+++ src/app/cms/components/sidenav-list/sidenav-list.component.ts
@@ -1,5 +1,5 @@
 import { Component, EventEmitter, OnInit, Output } from '@angular/core'
-import { routeNames } from '../../../cms-routing.module'
+import { routeNames } from '../../cms-routing.module'
 import { RouteName } from 'src/app/cms/cms.paths'
 
 @Component({
