+++ src/app/cms/root/cms.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core'
-import { AuthService } from '../../services/auth.service'
-import { DatabaseService } from '../../services/database.service'
+import { AuthService } from '../services/auth.service'
+import { DatabaseService } from '../services/database.service'
 
 @Component({
   selector: 'app-cms',
