+++ src/app/cms/routes/file-upload/components/progress-card/progress-card.component.scss
