+++ src/app/cms/routes/file-upload/components/upload-card/upload-card.component.scss
