+++ src/app/cms/routes/file-upload/file-upload.component.scss
