+++ src/app/cms/routes/login/login.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core'
 import { NgForm } from '@angular/forms'
-import { AuthService } from '../../../services/auth.service'
+import { AuthService } from '../../services/auth.service'
 
 @Component({
   selector: 'app-login',
